+++ src/main/java/com/kakao/cafe/service/MemberService.java
@@ -19,7 +19,7 @@ public class MemberService {
     }
 
     //회원가입
-    public Long join(Member member) {
+    public int join(Member member) {
         memberRepository.save(member);
         return member.getId();
     }
@@ -30,7 +30,7 @@ public class MemberService {
     }
 
     //회원 id에 해당하는 회원 name을 찾기
-    public Optional<Member> findOne(Long memberId) {
+    public Optional<Member> findOne(int memberId) {
         return memberRepository.findById(memberId);
     }
 }
